requirements.txt
Flask
Flask-CORS
numpy
tensorflow==2.17.0
pandas
scikit-learn
joblib
pickle-mixin
Pillow
gunicorn
gdown
